requirements.txt
ansible==2.9.27
apturl==0.5.2
asgiref==3.5.0
attrs==22.1.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
configparser==5.3.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
fasteners==0.14.1
filelock==3.8.0
frappe-bench==5.14.2
future==0.18.2
gitdb2==2.0.6
GitPython==2.1.15
gyp==0.1
h11==0.13.0
honcho==1.1.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
jeepney==0.7.1
Jinja2==3.0.3
kazam==1.4.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
linecache2==1.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
pickleDB==0.9.2
Pillow==9.0.1
platformdirs==2.5.2
pluggy==1.0.0
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.11.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==7.1.3
pytest-ansible==2.2.4
python-apt==2.3.0+ubuntu2.1
python-crontab==2.6.0
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pyusb==1.2.1
pyxdg==0.27
PyYAML==5.4.1
pyzk==0.9
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
semantic-version==2.8.5
six==1.16.0
smmap==5.0.0
smmap2==3.0.1
ssh-import-id==5.11
systemd-python==234
terminator==2.1.1
testfm==0.1.4
tomli==2.0.1
traceback2==1.4.0
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.15.0
virtualenv==20.16.5
wadllib==1.3.6
wsproto==1.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
